feat(choseyourcocktail): add the page navigation
