+++ src/components/ChooseYourCocktail.jsx
@@ -1,18 +1,26 @@
 import React from 'react';
-import { Link } from "react-router-dom";
+import { Link, useLocation } from "react-router-dom";
+import CocktailPreview from "./CocktailPreview";
 
 
-class ChooseYourCocktail extends React.Component {
-    render() {
-        return(
-            <div>
-                <div>
-                    <Link className="homeButton" to="/">X</Link>
-                </div>
-                WorkingProgress
-            </div>
-        )
-    }
+function useQuery() {
+  return new URLSearchParams(useLocation().search);
+}
+
+function ChooseYourCocktail({ cocktailList }) {
+  const query = useQuery();
+  const page = parseInt(query.get('page'), 10);
+  const pageSize = 3;
+  return (
+    <div>
+      <div>
+          <Link className="homeButton" to="/">X</Link>
+      </div>
+      <div>There is a cocktail list</div>
+      {cocktailList.slice((page - 1) * pageSize, page * pageSize).map((cocktail) => <CocktailPreview cocktail={cocktail} />)}
+      
+    </div>
+  );
 }
