+++ src/components/CocktailList.jsx
@@ -4,12 +4,13 @@ import CocktailPreview from './CocktailPreview';
 function CocktailList({ list, errorMessage, errorShow, manageError }) {
   return (
     <div>
+      
       <div className={errorShow === true ? 'errorMessage errorModalBg' : 'hideDescription'}>
         error message {errorMessage}
         <button type="button" onClick={manageError}>ok</button>
       </div>
       <div>
-        { list.map((cocktail) => <CocktailPreview cocktail={cocktail} />) }
+        {list.map((cocktail) => <CocktailPreview cocktail={cocktail} />) }
       </div>
     </div>
   );
